refactor - move all but the functions from CreateDirectMessage to MembersModal
